test OpenAPI spec generation using --dry-run command line parameter

+++ src/pyload/core/__init__.py
@@ -376,18 +376,22 @@ class Core:
     def _generate_open_api_spec(self):
         from pyload.webui.app.api_docs.openapi_specification_generator import OpenAPISpecificationGenerator
 
-        last_index = __file__.rfind("src"+ os.sep + "pyload")
-        if last_index != -1:
-            base_path =  pathlib.Path(__file__[:last_index])
-            if spec_path := next((p for p in base_path.rglob('openapi.json')), None):
-                self.log.debug("Saving OpenAPI spec to: %s", spec_path)
-                openapi_spec = OpenAPISpecificationGenerator(api=self.api).generate_openapi_json()
-                with open(spec_path, 'w') as f:
-                    json.dump(openapi_spec, f, indent=2)
+        self.log.debug("Generating OpenAPI spec")
+        openapi_spec = OpenAPISpecificationGenerator(api=self.api).generate_openapi_json()
+        self.log.debug("OpenAPI spec has been generated")
+
+        if self._dry_run is False:
+            last_index = __file__.rfind("src"+ os.sep + "pyload")
+            if last_index != -1:
+                base_path =  pathlib.Path(__file__[:last_index])
+                if spec_path := next((p for p in base_path.rglob('openapi.json')), None):
+                    self.log.debug(f"Saving OpenAPI spec to: {spec_path}")
+                    with open(spec_path, 'w') as f:
+                        json.dump(openapi_spec, f, indent=2)
+                else:
+                    raise IOError("Unable to locate openapi.json file")
             else:
-                raise IOError("Unable to locate openapi.json file")
-        else:
-            raise IOError("Unable to locate pyLoad's project directory")
+                raise IOError("Unable to locate pyLoad's project directory")
 
     def start(self):
         try:
